requirements.txt
PyMySQL
PyYAML
adbutils
allure-pytest
allure-python-commons
beautifulsoup4
bs4
decorator
loguru
lxml
mysql-connector-python
packaging
pip
pytest
requests
uiautomator2
urllib3
weditor

